% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{add_arch_plot}
\alias{add_arch_plot}
\title{add_arch_plot}
\usage{
add_arch_plot(
  hic_mat,
  qgr,
  src_gr = qgr,
  min_arch_dist = 1 * 10^6,
  n_arch = 50,
  p = NULL
)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
add_arch_plot
}
