% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{add_matrix_plot}
\alias{add_matrix_plot}
\title{Title}
\usage{
add_matrix_plot(
  hic_mat,
  qgr,
  tile_type = c("diamond", "hex")[1],
  max_dist = 10 * 10^6,
  point_size = 3,
  text_size = 4,
  max_fill = NULL,
  hmap_colors = c("lightgray", "steelblue", "darkblue", "red", "red"),
  minmax2col = c(min = "orange", max = "green"),
  show_insulation_range = T,
  fill_limits = NULL,
  show_min = T,
  show_max = F,
  p = NULL
)
}
\arguments{
\item{p}{}
}
\description{
Title
}
