% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_matrix_helpers.R
\name{apply_diagonal_zscore}
\alias{apply_diagonal_zscore}
\title{apply_diagonal_zscore}
\usage{
apply_diagonal_zscore(hic_mat, include_hetero_chr = FALSE)
}
\arguments{
\item{hic_mat}{}
}
\value{

}
\description{
calculates z-score of log2 transformed data
z-scores are segmented by chromosome and by distance from diagonal
returns recalculated HiC_matrix_wInsulation
}
