% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_matrix_helpers.R
\name{decrease_resolution}
\alias{decrease_resolution}
\title{decrease_resolution}
\usage{
decrease_resolution(hic_mat, pool_factor, calc_insulation = F)
}
\arguments{
\item{calc_insulation}{}
}
\description{
decrease_resolution
}
