% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_directionalityIndex.R
\name{directionality_of_index}
\alias{directionality_of_index}
\title{DI calculation for single index}
\usage{
directionality_of_index(hic_matrix, pos_index, n_bins)
}
\arguments{
\item{n_bins}{}
}
\description{
calculates the directionality index of
pos_index position using n_bins in both directions
}
