% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_matrix_helpers.R
\name{fetch_hic_byGRanges}
\alias{fetch_hic_byGRanges}
\title{fetch_hic_byGRanges}
\usage{
fetch_hic_byGRanges(hic_mat, gr1, gr2 = gr1, ext = 0)
}
\arguments{
\item{hic_mat}{a valid HiC_matrix object}

\item{gr1}{GRanges, specifies range of i indexes}

\item{gr2}{Granges, specifies range of j indexes, default is same as gr1}

\item{ext}{integer, extends gr1 and gr2 by set amount, default is 0}
}
\value{
data.table correspoding to gr1xgr2 region of hic_mat
}
\description{
Retrieves rectangular region of a HiC_matrix.
}
