% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bigInteract.R
\name{fetchBigInteract}
\alias{fetchBigInteract}
\title{fetchBigInteract}
\usage{
fetchBigInteract(bif, qgr, qgr2 = NULL, silent = FALSE)
}
\arguments{
\item{silent}{if TRUE, no warnings will be shown. default is FALSE.}
}
\value{

}
\description{
fetchBigInteract
}
