% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{ggplot_ref}
\alias{ggplot_ref}
\title{Title}
\usage{
ggplot_ref(
  gene_gr,
  qgr,
  label_method = c("text", "label", "none")[1],
  gene_types = "protein_coding",
  highlight_gene_names = "",
  highlight_gene_color = "red",
  olap_extension = 0.03,
  top_spacer = 0.8,
  strand_colors = c(`+` = "gray0", `-` = "gray40"),
  arrow_override = arrow(ends = "last", length = unit(0.045, "npc")),
  line_size = 2,
  line_end = c("round", "butt", "square")[2],
  text_hjust = 0,
  text_vjust = -1.5,
  text_angle = 30,
  text_size = 3.5,
  text_x_relative = 0.5,
  text_y_relative = 0.1
)
}
\arguments{
\item{text_y_relative}{}
}
\description{
Title
}
