% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HiC_matrix.from_hic.R
\name{HiC_matrix.from_hic}
\alias{HiC_matrix.from_hic}
\title{Load a HiC_matrix from a .hic file.}
\usage{
HiC_matrix.from_hic(hic_f, query_gr, bin_size = NULL)
}
\arguments{
\item{hic_f}{a .hic file}

\item{query_gr}{a query GRanges}

\item{bin_size}{the bin size to retrieve. Use strawr::readHicBpResolutions to find valid resoltions.}
}
\description{
This HiC_matrix is only a small slice of complete matrix therefore stats like insulation are not valid.
}
