% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HiC_matrix.R
\name{HiC_matrix}
\alias{HiC_matrix}
\title{HiC_matrix constructor}
\usage{
HiC_matrix(matrix_file, regions_file = NULL, parameters = NULL)
}
\arguments{
\item{matrix_file}{HiC-Pro matrix file}

\item{regions_file}{HiC-Pro region bed file matching matrix file}

\item{parameters}{HiC_parameters object}
}
\value{

}
\description{
HiC_matrix constructor
}
