% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HiC_matrix_wDirectionality.R
\name{HiC_matrix_wDirectionality}
\alias{HiC_matrix_wDirectionality}
\title{HiC_matrix_wDirectionality constructor}
\usage{
HiC_matrix_wDirectionality(hic_matrix)
}
\arguments{
\item{hic_matrix}{a loaded HiC_matrix}
}
\description{
HiC_matrix_wDirectionality constructor
}
