% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HiC_matrix_wInsulation.R
\name{HiC_matrix_wInsulation}
\alias{HiC_matrix_wInsulation}
\title{HiC_matrix_wInsulation constructor}
\usage{
HiC_matrix_wInsulation(hic_matrix)
}
\arguments{
\item{hic_matrix}{a loaded HiC_matrix}
}
\description{
HiC_matrix_wInsulation constructor
}
