% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_combined_hic}
\alias{plot_combined_hic}
\title{Title}
\usage{
plot_combined_hic(
  hic_list,
  runx_bws,
  ctcf_bws,
  exons_gr = NULL,
  max_dist = NULL,
  hic_names = names(hic_list),
  target_gene,
  extension_size_bp,
  output_prefix = "combined_plots",
  n_arch = 50,
  min_arch_dist = 10 * 10^5,
  max_fill = NULL,
  hmap_colors = c("lightgray", "steelblue", "darkblue", "red", "red")
)
}
\arguments{
\item{hmap_colors}{}
}
\value{

}
\description{
Title
}
