% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_upperMatrix}
\alias{plot_upperMatrix}
\title{Title}
\usage{
plot_upperMatrix(
  hic_mat,
  chr,
  start,
  end,
  hmap_colors = c("lightgray", "steelblue", "darkblue", "red", "red")
)
}
\arguments{
\item{hmap_colors}{}
}
\value{

}
\description{
Title
}
