% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_upperMatrix_with_insulation}
\alias{plot_upperMatrix_with_insulation}
\title{Title}
\usage{
plot_upperMatrix_with_insulation(
  hic_mat,
  chr,
  start,
  end,
  tile_type = c("diamond", "hex")[1],
  show_plot = T,
  main_title = NULL,
  max_dist = 10 * 10^6,
  point_size = 3.5,
  max_fill = NULL,
  hmap_colors = c("lightgray", "steelblue", "darkblue", "red", "red"),
  show_insulation_range = T,
  show_minmax = T
)
}
\arguments{
\item{show_minmax}{}
}
\value{

}
\description{
Title
}
