% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_matrix_helpers.R
\name{subset_dt}
\alias{subset_dt}
\title{Subset data.table by i,j indexes}
\usage{
subset_dt(dt, i_range, j_range, fill_square = FALSE)
}
\arguments{
\item{dt}{a data.table containing i, j, and val}

\item{i_range}{integer, range of i indexes to retrieve}

\item{j_range}{integer, range of j indexes to retrieve}

\item{fill_square}{logical, should region spanning below diagonal be filled in?}
}
\value{
subset of dt within ranges
}
\description{
efficiently subset data.table as one would a subrange of a matrix
}
