% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_upperMatrix}
\alias{plot_upperMatrix}
\title{plot_upperMatrix}
\usage{
plot_upperMatrix(
  hic_mat,
  chr,
  start,
  end,
  hmap_colors = c("lightgray", "steelblue", "darkblue", "red", "red"),
  fill_limits = c(NA, NA),
  na.value = hmap_colors[length(hmap_colors)]
)
}
\arguments{
\item{hmap_colors}{}
}
\description{
plot_upperMatrix
}
